Compartimentação
Modularização. Coloquei a HomeScreen num arquivo e o Onboarding em outro. Obrigado, Código Limpo.
